requirements.txt
Flask
gunicorn
SQLAlchemy
psycopg2-binary
python-dotenv